# performance improvements backport from the pravega driver