--- a/build.gradle
+++ b/build.gradle
@@ -29,7 +29,7 @@ targetCompatibility = 11
 
 ext {
 	depVersion = [
-		commonsJava              : "2.3.5",
+		commonsJava              : "2.3.6",
 		confuse                  : "1.2.1",
 		confuseIoYaml            : "1.0.0",
 		disruptor                : "3.4.2",
@@ -37,8 +37,8 @@ ext {
 		javaxActivation			 : "1.1",
 		kafka					 : "2.2.0", 
 		log4j                    : "2.8.2",
-		mongooseBase             : "4.2.6",
-		mongooseStorageDriverCoop: "4.2.6",
+		mongooseBase             : "4.2.11",
+		mongooseStorageDriverPreempt: "4.2.13",
 		scala                    : "2.12.6",
 	]
 	rootPkg = "com.emc.mongoose.storage.driver.kafka"
@@ -71,7 +71,7 @@ dependencies {
 	compileOnly(
 		"org.projectlombok:lombok:1.18.6",
 		"com.github.emc-mongoose:mongoose-base:${depVersion.mongooseBase}",
-		"com.github.emc-mongoose:mongoose-storage-driver-coop:${depVersion.mongooseStorageDriverCoop}",
+		"com.github.emc-mongoose:mongoose-storage-driver-preempt:${depVersion.mongooseStorageDriverPreempt}",
 		"com.github.akurilov:confuse:${depVersion.confuse}",
 		"com.github.akurilov:confuse-io-yaml:${depVersion.confuseIoYaml}",
 		"com.github.akurilov:java-commons:${depVersion.commonsJava}",