--- a/src/main/resources/config-schema-storage-driver-kafka.yaml
+++ b/src/main/resources/config-schema-storage-driver-kafka.yaml
@@ -1,7 +1,7 @@
 ---
 load:
   op:
-    timeoutMillis: int
+    timeoutMillis: long
 storage:
   driver:
     create:
@@ -11,8 +11,6 @@ storage:
     read:
       timeoutMillis: long
     request:
-      size: int
-    batch:
       size: int
     buffer:
       memory: long