- Inherit the storage driver based on the thread pool executor (aka preempt)
- Introduce batch mode
- Thread local producer for the contention avoiding
- Fixes

Results: write rate is more than 500,000 per second smallest records per second (locally)
```
java -jar ~/work/mongoose-base/build/libs/mongoose-4.2.11.jar --storage-driver-type=kafka --item-data-size=1 --storage-driver-threads=8 --storage-driver-limit-concurrency=0 --load-batch-size=10000 --storage-driver-limit-queue-input=100
Picked up _JAVA_OPTIONS: -Dawt.useSystemAAFontSettings=on
################################################## mongoose v 4.2.11 ##################################################
2019-05-25T20:18:26,684 I                                main                           Available/installed extensions:
        Load --------------------------> com.emc.mongoose.base.load.step.linear.LinearLoadStepExtension
        dummy-mock --------------------> com.emc.mongoose.base.storage.driver.mock.DummyStorageDriverMockExtension
        PipelineLoad ------------------> com.emc.mongoose.load.step.pipeline.PipelineLoadStepExtension
        coop-mock ---------------------> com.emc.mongoose.storage.driver.coop.mock.CoopStorageDriverMockExtension
        pravega -----------------------> com.emc.mongoose.storage.driver.pravega.PravegaStorageDriverExtension
        kafka -------------------------> com.emc.mongoose.storage.driver.kafka.KafkaStorageDriverExtension
        preempt-mock ------------------> com.emc.mongoose.storage.driver.preempt.mock.PreemptStorageDriverMockExtension

Warning: Nashorn engine is planned to be removed from a future JDK release
2019-05-25T20:18:27,186 I                                main                           Using the "Oracle Nashorn" scenario engine
2019-05-25T20:18:27,840 I StorageDriver                  main                           linear_20190525.201827.793: creating the storage driver instance for the type "kafka"
2019-05-25T20:18:27,922 I StorageDriver                  main                           linear_20190525.201827.793: creating the storage driver instance for the type "kafka"
2019-05-25T20:18:28,144 I LinearLoadStepClient           main                           linear_20190525.201827.793: load step client started, additional nodes: []
------------------------------------------------------------------------------------------------------------------------
 Step Id  | Timestamp  |  Op  |     Concurrency     |       Count       | Step  |   Last Rate    |  Mean    |   Mean    
 (last 10 |            | type |---------------------|-------------------| Time  |----------------| Latency  | Duration  
 symbols) |yyMMddHHmmss|      | Current  |   Mean   |   Success  |Failed|  [s]  | [op/s] |[MB/s] |  [us]    |   [us]    
----------|------------|------|----------|----------|------------|------|-------|--------|-------|----------|-----------
201827.793|190525201828|CREATE|         1|1.0       |           0|     0|0.006  |0.0     |0.0    |         0|          0
201827.793|190525201838|CREATE|         8|7.52222222|     1052872|     0|10.122 |189018.1|0.18026|     16025|      16073
201827.793|190525201848|CREATE|         8|7.75144508|     3028730|     0|20.169 |216303.0|0.20628|     13349|      13386
201827.793|190525201858|CREATE|         7|6.53584905|     7188741|     0|30.169 |498844.2|0.47573|     17264|      17283
201827.793|190525201908|CREATE|         5|5.90909090|    12558739|     0|40.169 |564473.2|0.53832|     18448|      18461
201827.793|190525201918|CREATE|         5|5.55869565|    18041722|     0|50.169 |566909.2|0.54064|     18200|      18210
201827.793|190525201928|CREATE|         3|5.29316546|    23595407|     0|60.169 |594825.9|0.56726|     17793|      17802
201827.793|190525201938|CREATE|         2|5.21625766|    29227289|     0|70.169 |577748.1|0.55098|     19054|      19061
201827.793|190525201948|CREATE|         5|5.26203208|    34911015|     0|80.169 |571631.9|0.54515|     21270|      21277
201827.793|190525201958|CREATE|         2|5.15147928|    40503180|     0|90.169 |580512.2|0.55361|     20954|      20961
201827.793|190525202008|CREATE|         1|5.09013785|    46115521|     0|100.18 |563279.9|0.53718|     21387|      21393
201827.793|190525202018|CREATE|         4|5.02694898|    51653623|     0|110.186|561463.8|0.53545|     21297|      21303
201827.793|190525202028|CREATE|         5|4.98328935|    57237600|     0|120.186|563317.4|0.53722|     21405|      21411
201827.793|190525202038|CREATE|         4|4.91734197|    62818058|     0|130.186|574046.8|0.54745|     21222|      21227
```